+++ database/migrations/2023_10_13_145300_create_attendances_table.php
@@ -11,19 +11,21 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('attendances', function (Blueprint $table) {
-            $table->id();
-            $table->date('date');
-            $table->unsignedBigInteger('emp_id');
-            $table->integer('hours')->default(0);
-            $table->integer('user_id');
-            $table->timestamps();
-
-            $table->unique(['date', 'emp_id']);
-
-            // Foreign key constraint for emp_id
-            $table->foreign('emp_id')->references('id')->on('employees')->onDelete('cascade');
-        });
+        if(! Schema::hasTable('attendances')) {
+            Schema::create('attendances', function (Blueprint $table) {
+                $table->id();
+                $table->date('date');
+                $table->unsignedBigInteger('emp_id');
+                $table->integer('hours')->default(0);
+                $table->integer('user_id');
+                $table->timestamps();
+    
+                $table->unique(['date', 'emp_id']);
+    
+                // Foreign key constraint for emp_id
+                $table->foreign('emp_id')->references('id')->on('employees')->onDelete('cascade');
+            });
+        }
     }
 
     /**
