+++ database/migrations/2023_11_14_004509_update_purchase_add_total_due_delete_asof_date_due.php
@@ -11,10 +11,12 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('purchases', function (Blueprint $table) {
-            $table->float('total_due', 14, 2)->default(0)->after('total');
-            $table->dropColumn('asof_date_due');
-        });
+        if(! Schema::hasColumns('purchases', ['total_due'])) {
+            Schema::table('purchases', function (Blueprint $table) {
+                $table->float('total_due', 14, 2)->default(0)->after('total');
+                $table->dropColumn('asof_date_due');
+            });
+        }
     }
 
     /**
